Fix size_t is cast to int (bug when len == SIZE_MAX)

+++ ft_strndup.c
@@ -6,7 +6,7 @@
 /*   By: thugo <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/22 18:20:45 by thugo             #+#    #+#             */
-/*   Updated: 2016/11/23 11:04:08 by thugo            ###   ########.fr       */
+/*   Updated: 2017/01/24 16:14:44 by thugo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,15 +16,15 @@
 char	*ft_strndup(const char *s, size_t len)
 {
 	char	*new;
-	int		i;
+	size_t	i;
 	size_t	size;
 
-	size = (int)ft_strlen(s);
+	size = ft_strlen(s);
 	new = (char *)malloc(sizeof(char) * (size > len ? len : size) + 1);
 	if (new == NULL)
 		return (NULL);
 	i = 0;
-	while (s[i] != '\0' && i < (int)len)
+	while (s[i] != '\0' && i < len)
 	{
 		new[i] = s[i];
 		i++;
